refactor: move GridBox out of TeamSection
